requirements.txt
crewai==0.30.11
crewai-tools==0.1.7
streamlit==1.35.0
python-dotenv==1.0.1
google-generativeai==0.5.4
tavily-python==0.3.3